src/main/java/ua/hrynko/BL/Cell.java
package ua.hrynko.BL;

import java.util.Objects;

/**
 * Used to represent cell on a field
 */
public class Cell {
    private int id;
    private boolean isHit;
    private boolean isBusy;
    private int x;
    private int y;

    public Cell(int id, int x, int y) {
        this.id = id;
        this.x = x;
        this.y = y;
    }

    public int getId() {
        return id;
    }

    public boolean isBusy() {
        return isBusy;
    }

    public void setBusy(boolean busy) {
        isBusy = busy;
    }

    public void toggleBusy() {
        isBusy = !isBusy;
    }

    public boolean isHit() {
        return isHit;
    }

    public void setHit(boolean hit) {
        isHit = hit;
    }

    @Override
    public boolean equals(Object o) {
        if (this == o) return true;
        if (o == null || getClass() != o.getClass()) return false;
        Cell cell = (Cell) o;
        return id == cell.id;
    }

    @Override
    public int hashCode() {
        return Objects.hash(id);
    }
}
